requirements.txt
Django==5.2.4
Pillow
django-widget-tweaks
requests
python-dotenv